Change profile initialization processing timing
